feat(print_results): テストの出力順序を更新
